tweaks paths
The debug and srpms are signed, so we can safely turn on the gpgcheck. 

Also, the debugs are dropped into a common directory for all the storage sig ( this is for all SIGs ), rather than split them by project name. The assumption being that there will never be the same rpm ( name-version-release.arch ) from two different projects in the same SIG ( they should just tag each others content as needed ).